+++ NEWS.md
@@ -1,3 +1,9 @@
+# usedthese 0.3.0
+
+* Respects `include.only` and `exclude` arguments specified in `library()`
+* Small performance improvement with dplyr 1.1 and tidyr 1.3
+* Remove suggests for meta-packages tidyverse and fpp3
+
 # usedthese 0.2.0
 
 * Support use of the conflicted package